fix linking error due to global variables defined in *.h
When building, there are many linking errors:
/usr/bin/ld: CMakeFiles/zyncore.dir/zyncontrol_vx.c.o:(.bss+0x208):
        multiple definition of `midi_learning_mode';
        CMakeFiles/zyncore.dir/zyncore.c.o:(.bss+0x4): first defined here
/usr/bin/ld: CMakeFiles/zyncore.dir/zyncontrol_vx.c.o:(.bss+0x20c):
        multiple definition of `zmops';
        CMakeFiles/zyncore.dir/zyncore.c.o:(.bss+0x8): first defined here

Those are caused by many variables defined in header files, this led to
have many .c files to have a definition of the same variable.

Many of those variables which are not shared have been moved to the
corresponding .c file. Those which are shared have been declared as
extern.